requirements.txt
Django==4.0
djangorestframework==3.13
django-redis==5.2.0
psycopg2>=2.8
redis==4.2.2

celery==5.2.6
django-celery-beat>=2.0
django-celery-results>=2.0

stripe>=2.0
python-decouple>=3